+++ src/codegen/php/mod.rs
@@ -7,7 +7,7 @@ use super::Transpilable;
 impl Transpilable for PhpAst<'_> {
     fn transpile(&self) -> String {
         let mut fragments = vec![String::from("<?php\n")];
-        
+
         for statement in self.statements.iter() {
             fragments.push(statement.transpile());
         }
@@ -20,7 +20,8 @@ impl Transpilable for PhpStatement<'_> {
     fn transpile(&self) -> String {
         match self {
             PhpStatement::PhpEchoStatement(expr_list) => {
-                let expressions_vec = expr_list.expressions
+                let expressions_vec = expr_list
+                    .expressions
                     .iter()
                     .map(|e| e.transpile())
                     .collect::<Vec<_>>();
@@ -46,5 +47,3 @@ impl Transpilable for PhpExpression<'_> {
         }
     }
 }
-
-