+++ CHANGELOG.md
@@ -26,7 +26,7 @@
 ## v0.0.15
 
 - [x] Multiline comments
-- [ ] Nested multiline comments
+- [x] Nested multiline comments
 - [ ] Include comments in the AST
 - [ ] Replace all panics with actual errors
 - [ ] Remove all old codegen